@@ -7,7 +7,7 @@
   <PropertyGroup>
     <RepoRoot>$(MsBuildThisFileDirectory)</RepoRoot>
     <LangVersion>latest</LangVersion>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <CodeAnalysisRuleSet>$(MsBuildThisFileDirectory)Analyzers/rulesets.xml</CodeAnalysisRuleSet>
   </PropertyGroup>
 
