@@ -113,7 +113,7 @@ public class ExamGroupBusinessService : AdministrationOperationService<ExamGroup
         await this.AddUsersToExamGroup(examGroup, users);
 
         var externalUsernames = usernames
-            .Except(users.Select(u => u.UserName), StringComparer.OrdinalIgnoreCase)
+            .Except(users.Select(u => u.UserName!), StringComparer.OrdinalIgnoreCase)
             .ToList();
 
         if (externalUsernames.Any())
