@@ -79,7 +79,7 @@ namespace OJS.Services.Ui.Data.Implementations
 
             if (existingScore == null)
             {
-                await this.AddBySubmissionByUsernameAndIsOfficial(submission, participant.UserName, participant.IsOfficial, participant.Participant);
+                await this.AddBySubmissionByUsernameAndIsOfficial(submission, participant.UserName!, participant.IsOfficial, participant.Participant);
             }
             else
             {
