@@ -34,5 +34,10 @@ public class ContestParticipationServiceModel : IMapExplicitly
             .ForMember(d => d.IsInvalidated, opt => opt.MapFrom(s => s.IsInvalidated))
             .ForMember(d => d.Contest, opt => opt.Ignore())
             .ForMember(d => d.EndDateTimeForParticipantOrContest, opt => opt.Ignore())
-            .ForAllOtherMembers(opt => opt.Ignore());
+            .ForMember(m => m.IsRegisteredParticipant, opt => opt.Ignore())
+            .ForMember(m => m.IsActiveParticipant, opt => opt.Ignore())
+            .ForMember(m => m.ParticipantId, opt => opt.Ignore())
+            .ForMember(m => m.LastSubmissionTime, opt => opt.Ignore())
+            .ForMember(m => m.UserSubmissionsTimeLimit, opt => opt.Ignore())
+            .ForMember(m => m.ParticipantsCount, opt => opt.Ignore());
 }
